[desktop] Fix flakiness in reading zip files
I'm not sure what was the issue in the existing code, but I happened to chance
on a setup that reproduced the flakiness that some customers have reported (that
reading the zips sometimes fails). There wasn't anything specific in the setup -
I was reading a 50 MB zip file, a file which I'd read multiple times before,
except this time it seemed to invariably result in failures during read.

Replacing the node stream to web stream conversion with this new approach fixes
the flakiness, at least in the reproducible scenario that I was encountering.