Start with a idle state
Not sure why it was true - there is a possibility that this was intentional, but
I can't think why. The reason for changing it is to fix our "isForced" logic
(otherwise the non-file-related sync doesn't run on app start), without
introducing _another_ flag to track if the sync was initiated from a gallery
useEffect or by the preUploadSync.