Add the ability to show the Ente Auth window on macOS after it was closed.
Currently, nothing happens when the user closes the Ente Auth window on macOS and then clicks on the Ente Auth icon in the dock. This commit changes this behaviour and brings it in-line with other macOS applications by showing the main application window when the user clicks on the dock icon.